fix: add explicit casts for implicit narrowing conversions (alerts #131-#136)
- Request.java: cast long to int in postSize += (int) part.getSize()
- Http2UpgradeHandler.java: explicit byte cast for FLAG_END_OF_HEADERS addition
- MultipartStream.java: cast long to int in head += (int) res
- WsRemoteEndpointImplBase.java: explicit byte cast for rsv shift addition
- Http2TestBase.java: explicit byte casts for 0x08 flag additions (2 sites)

+++ java/org/apache/catalina/connector/Request.java
@@ -2862,7 +2862,7 @@ public class Request implements HttpServletRequest {
                             // Equals sign
                             postSize++;
                             // Value length
-                            postSize += part.getSize();
+                            postSize += (int) part.getSize();
                             // Value separator
                             postSize++;
                             if (postSize > maxPostSize) {
